backend/requirements.txt
# GPU-Optimized Audio Processing & AI for VoIP Scam Detection
# =============================================================

# Web Framework (Optional - only if using main.py instead of simple_processor.py)
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6

# GPU-Optimized PyTorch (CUDA support)
# Install with: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchaudio>=2.1.0

# GPU-Optimized TensorFlow (CUDA support)
tensorflow[and-cuda]>=2.15.0
tensorflow-hub>=0.15.0

# Audio Processing & Speech Recognition
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Machine Learning & Transformers (GPU acceleration)
transformers>=4.35.0
accelerate>=0.24.0

# Speaker Diarization (GPU optimized)
pyannote.audio>=3.1.0

# Scientific Computing (optimized)
numpy>=1.24.0
scipy>=1.10.0

# File System Monitoring
watchdog>=3.0.0

# Audio Format Conversion
ffmpeg-python>=0.2.0

# Process Management
psutil>=5.9.0

# Additional ML tools
pandas>=2.0.0
scikit-learn>=1.3.0

# GPU Memory Management
gpustat>=1.1.0

# CUDA toolkit support (if needed)
# Note: Install CUDA 12.1+ separately for best performance
